requirements.txt
Django==5.0.3
psycopg2-binary==2.9.9
PyPDF2==3.0.1
google-generativeai==0.5.0
python-dotenv==1.0.1
gunicorn==21.2.0
